+++ Gyrocam/Gyrocam/Geometry.h
@@ -6,9 +6,9 @@ using namespace std;
 
 namespace gyrocam
 {
-	Point3f lineThroughPoints(Point3f p1, Point3f p2);
-	Point3f intersection(Point3f l1, Point3f l2);
-	double incidence(Point3f p1, Point3f p2);
-	bool isInfinitePoint(Point3f p);
-	void normalizeZ(Point3f &p);
+	Point3d lineThroughPoints(Point3d p1, Point3d p2);
+	Point3d intersection(Point3d l1, Point3d l2);
+	double incidence(Point3d p1, Point3d p2);
+	bool isInfinitePoint(Point3d p);
+	void normalizeZ(Point3d &p);
 }
