+++ Gyrocam/Gyrocam/Shared.h
@@ -9,10 +9,10 @@ using namespace std;
 
 namespace gyrocam
 {
-	Point3f toProjective(Vec4i segment, bool first = true);
-	Point3f toProjective(float x, float y);
-	double norm12(Point3f p);
-	bool isIncident(Point3f vp, LineSegment segment);
+	Point3d toProjective(Vec4i segment, bool first = true);
+	Point3d toProjective(double x, double y);
+	double norm12(Point3d p);
+	bool isIncident(Point3d vp, LineSegment segment);
 	void setRow(Mat a, int i, Mat col);
 	
 	int iterateOnSet(set<int> &base, set<int>::iterator &it, int shift);
